backend/requirements.txt
# FastAPI Core
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
